pyproject.toml
[tool.poetry]
name = "dailydo-todo-app"
version = "1.0.0"
description = ""
authors = ["Saad-Kamran-2006 <saadkamran6ft@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.5"
fastapi = "^0.112.0"
httpx = "^0.27.0"
sqlmodel = "^0.0.21"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pytest = "^8.3.2"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
types-passlib = "^1.7.7.20240327"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-python-jose = "^3.3.4.20240106"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
